+++ src/main/java/chess/dto/BoardDto.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.dto;
 
 import java.util.List;
