+++ src/main/java/chess/model/Board.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.model;
 
 import chess.model.piece.Bishop;
 import chess.model.piece.Color;
@@ -16,6 +16,7 @@ import chess.model.status.End;
 import chess.model.status.Ready;
 import chess.model.status.Running;
 import chess.model.status.Status;
+import chess.model.util.ScoreResult;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
