+++ src/main/java/chess/model/piece/Direction.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.model.piece;
 
 import java.util.List;
